\name{seu2seuList}
\alias{seu2seuList}
\title{Seurat object to seuratList object}
\description{
  Transfer Seurat object to a Seurat list object for preparation for \link{PRECAST} model fitting. see our [PRECAST package website](https://feiyoung.github.io/PRECAST/index.html) for more usage of \link{PRECAST}.
}
\usage{
  seu2seuList(seu, batch, verbose=TRUE)
}
\arguments{
 \item{spe}{a Seurat object with multiple data bathes information. It must at least contain the batch(sampel) id in the meta.data (i.e., sample_id), the raw gene count expression matrix ans sptial coordinates. And the batch must be specified (i.e., batch='sample_id').}
  \item{batch}{a  string to specify the data batch field in the meta.data of seu. }
  \item{verbose}{an optional logical value, whether output the information.}
}

\value{
  Return a  list of multiple Seurat objects, where for each Seurat object, the spatial coordinates information is saved in the metadata of Seurat, named "row" and "col".
}

\author{
  Wei Liu
}
\note{
  nothing
}
\seealso{
  None
}

\examples{
  suppressPackageStartupMessages( library(SpatialExperiment))
  dir <- system.file(
    file.path("extdata", "10xVisium", "section1"),
    package = "SpatialExperiment")

  # read in counts
  fnm <- file.path(dir, "raw_feature_bc_matrix")
  sce <- DropletUtils::read10xCounts(fnm)

  # read in image data
  img <- readImgData(
    path = file.path(dir, "spatial"),
    sample_id="foo")

  # read in spatial coordinates
  fnm <- file.path(dir, "spatial", "tissue_positions_list.csv")
  xyz <- read.csv(fnm, header = FALSE,
                  col.names = c(
                    "barcode", "in_tissue", "array_row", "array_col",
                    "pxl_row_in_fullres", "pxl_col_in_fullres"))

  # construct observation & feature metadata
  rd <- S4Vectors::DataFrame(
    symbol = rowData(sce)$Symbol)

  # construct 'SpatialExperiment'
  (spe <- SpatialExperiment(
    assays = list(counts = assay(sce)),
    colData = colData(sce), rowData = rd, imgData = img,
    spatialData=DataFrame(xyz),
    spatialCoordsNames=c("pxl_col_in_fullres", "pxl_row_in_fullres"),
    sample_id="foo"))

  colnames(spe) <- paste0("spot", 1:ncol(spe))


  seu <- spe2seurat(spe)
  head(seu)
  seu2 <- seu
  seu2$sample_id <- paste0(seu2$sample_id, "2")
  library(SeuratObject)
  seu2 <- RenameCells(seu2,  paste0(colnames(seu2), "2") )
  seu_all <- merge(seu, seu2)
  head(seu_all@meta.data)
  table(seu_all$sample_id)
  seuList <- seu2seuList(seu_all, batch='sample_id')
  seuList

  ## Create PRECAST object
  library(PRECAST)
  Pobject <- CreatePRECASTObject(seuList=seuList, selectGenesMethod = "HVGs",
                               premin.spots = 0, premin.features = 0,
                               postmin.features = 0, postmin.spots = 0, verbose = F)
}

