\name{spe2seuratList}
\alias{spe2seuratList}
\title{SpatialExperiment object to Seurat list object}
\description{
  Transfer SpatialExperiment object to a Seurat list object for preparation for \link{PRECAST} model fitting. see our [PRECAST package website](https://feiyoung.github.io/PRECAST/index.html) for more usage of \link{PRECAST}.
}
\usage{
  spe2seuratList(spe, batch=NULL, verbose=TRUE)
}
\arguments{
  \item{spe}{a SpatialExperiment object or a list consisting of SpatialExperiment objects. If spe is a SpatialExperiment object, it must at least contain the batch(sampel) id in the colData (i.e., sample_id), the raw gene count expression matrix and spatial coordinates. And the batch must be specified (i.e., batch='sample_id'). If spe is a list consisting of multiple SpatialExperiment objects, then each object represents a data batch.}
  \item{batch}{a optional argument, NULL or a string. Only if spe is a list, batch can be NULL. }
  \item{verbose}{an optional logical value, whether output the information.}
}

\value{
  Return a list consisting of Seurat objects, where the spatial coordinates information is saved in the metadata of Seurat, named "row" and "col".
}

\author{
  Wei Liu
}
\note{
  nothing
}
\seealso{
  None
}

\examples{

  suppressPackageStartupMessages( library(SpatialExperiment))
  dir <- system.file(
    file.path("extdata", "10xVisium", "section1"),
    package = "SpatialExperiment")

  # read in counts
  fnm <- file.path(dir, "raw_feature_bc_matrix")
  sce <- DropletUtils::read10xCounts(fnm)

  # read in image data
  img <- readImgData(
    path = file.path(dir, "spatial"),
    sample_id="foo")

  # read in spatial coordinates
  fnm <- file.path(dir, "spatial", "tissue_positions_list.csv")
  xyz <- read.csv(fnm, header = FALSE,
                  col.names = c(
                    "barcode", "in_tissue", "array_row", "array_col",
                    "pxl_row_in_fullres", "pxl_col_in_fullres"))

  # construct observation & feature metadata
  rd <- S4Vectors::DataFrame(
    symbol = rowData(sce)$Symbol)

  # construct 'SpatialExperiment'
  (spe <- SpatialExperiment(
    assays = list(counts = assay(sce)),
    colData = colData(sce), rowData = rd, imgData = img,
    spatialData=DataFrame(xyz),
    spatialCoordsNames=c("pxl_col_in_fullres", "pxl_row_in_fullres"),
    sample_id="foo"))
  colnames(spe) <- paste0("spot", 1:ncol(spe))


  ## Transfer a SpatialExperiment to a seuList
  colData(spe)$batch_id <- rep(c("a", "b"), each=25)
  seuList1 <- spe2seuratList(spe, batch = 'batch_id')
  seuList1
  ## Transfer a list of SpatialExperiment to a seuList
  seuList2 <- spe2seuratList(list(spe, spe))
  seuList2

  ## Create PRECAST object
  library(PRECAST)
  Pobject <- CreatePRECASTObject(seuList=seuList2, selectGenesMethod = "HVGs",
                               premin.spots = 0, premin.features = 0,
                               postmin.features = 0, postmin.spots = 0, verbose = F)

}

