\name{spe2seurat}
\alias{spe2seurat}
\title{SpatialExperiment object to Seurat object}
\description{
  Transfer SpatialExperiment object to a Seurat object for preparation for \link{DR.SC} model fitting.
}
\usage{
  spe2seurat(spe, verbose= TRUE)
}
\arguments{
  \item{spe}{a SpatialExperiment object, at least including the raw gene count expression matrix ans sptial coordinates.}
  \item{verbose}{an optional logical value, whether output the information.}
}

\value{
  Return a  Seurat object, where the spatial coordinates information is saved in the metadata of Seurat, named "row" and "col".
}

\author{
  Wei Liu
}
\note{
  nothing
}
\seealso{
  None
}

\examples{

  dir <- system.file(
  file.path("extdata", "10xVisium", "section1"),
  package = "SpatialExperiment")

  # read in counts
  fnm <- file.path(dir, "raw_feature_bc_matrix")
  sce <- DropletUtils::read10xCounts(fnm)

  # read in image data
  img <- readImgData(
    path = file.path(dir, "spatial"),
    sample_id="foo")

  # read in spatial coordinates
  fnm <- file.path(dir, "spatial", "tissue_positions_list.csv")
  xyz <- read.csv(fnm, header = FALSE,
                  col.names = c(
                    "barcode", "in_tissue", "array_row", "array_col",
                    "pxl_row_in_fullres", "pxl_col_in_fullres"))

  # construct observation & feature metadata
  rd <- S4Vectors::DataFrame(
    symbol = rowData(sce)$Symbol)

  # construct 'SpatialExperiment'
  (spe <- SpatialExperiment(
    assays = list(counts = assay(sce)),
    colData = colData(sce), rowData = rd, imgData = img,
    spatialData=DataFrame(xyz),
    spatialCoordsNames=c("pxl_col_in_fullres", "pxl_row_in_fullres"),
    sample_id="foo"))

  colnames(spe) <- paste0("spot", 1:ncol(spe))


  seu <- spe2seurat(spe)
  head(seu)

  ## Fit DR-SC model
  library(DR.SC)
  seu <- NormalizeData(seu)
  seu <- FindVariableFeatures(seu)
  seu <- DR.SC(seu = seu, K=4)
}

