% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{srt.content}
\alias{srt.content}
\title{Retrieve Subtitle Text}
\usage{
srt.content(srt)
}
\arguments{
\item{srt}{vector. The srt file read by \code{\link[SRTtools]{srt.read}}.}
}
\description{
Retrieve all the subtitle text content from a srt file
}
\examples{
\dontrun{

srt <- srt.read("movie.srt", encoding = 'utf-8')

srt.content(srt)
}
}
\seealso{
\code{\link[SRTtools]{srt.read}}
}
