% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{srt.remove}
\alias{srt.remove}
\title{Remove dialog}
\usage{
srt.remove(srt, index)
}
\arguments{
\item{srt}{vector. The srt file read by \code{\link[SRTtools]{srt.read}}.}

\item{index}{integer. The index of unwanted dialog .}
}
\description{
Remove unwanted dialog from subtitles by specific index
}
\examples{
srt_path <- system.file("extdata", "movie.srt", package="SRTtools")
srt <- srt.read(srt_path, encoding = 'utf-8')
srt.remove(srt, index = 10)

}
\seealso{
\code{\link[SRTtools]{srt.read}}
}
