% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{srt.search}
\alias{srt.search}
\title{Search Index By KeyWord}
\usage{
srt.search(srt, key_word)
}
\arguments{
\item{srt}{vector. The srt file read by \code{\link[SRTtools]{srt.read}}.}

\item{key_word}{character. The key word want to be searched in subtitles.}
}
\description{
Return the subtitle index by specific keyword
}
\seealso{
\code{\link[SRTtools]{srt.read}}
}
