% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{srt.shift}
\alias{srt.shift}
\title{Re-synchronize srt file}
\usage{
srt.shift(srt, time_shifted)
}
\arguments{
\item{srt}{vector. The srt file read by \code{\link[SRTtools]{srt.read}}.}

\item{time_shifted}{numeric. The time that srt file want to be shifted (in seconds).}
}
\description{
Shift a srt file with specific time.
}
\examples{
\dontrun{

srt <- srt.read("movie.srt", encoding = 'utf-8')

# Postpone subtitles 3 seconds
srt <- srt.shift(srt, time_shifted = 3)

# Make subtitles 5 seconds earlier
srt <- srt.shift(srt, time_shifted = -5)
}
}
\seealso{
\code{\link[SRTtools]{srt.read}}
}
