% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SRTtools}
\alias{SRTtools}
\alias{SRTtools-package}
\title{Adjust srt file to get better experience in watching movie.}
\description{
Adjust time, style, and retrieve content for srt file.
}
\author{
Jim Chen
}
