% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{srt.search}
\alias{srt.search}
\title{Search Index By KeyWord}
\usage{
srt.search(srt, key_word)
}
\arguments{
\item{srt}{vector. The srt file read by \code{\link[SRTtools]{srt.read}}.}

\item{key_word}{character. The key word want to be searched in subtitles.}
}
\description{
Return the subtitle index by specific keyword
}
\examples{
srt_path <- system.file("extdata", "movie.srt", package="SRTtools")
srt <- srt.read(srt_path, encoding = 'utf-8')
srt.search(srt, key_word = "captain")

}
\seealso{
\code{\link[SRTtools]{srt.read}}
}
