% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SRTtools}
\alias{SRTtools}
\alias{SRTtools-package}
\title{Adjust srt file to get better experience in watching movie.}
\description{
Srt file is a common subtitle format for videos, it contains subtitle and when the subtitle showed.
This package is for ealign time of srt file, and also change color, style and position of subtitle in videos,
the srt file will be read as a vector into R, and can be write into srt file after modified using this package.
}
\author{
Jim Chen
}
